Damn Fine Commit
A bunch of assets for a first-person map with a reflective cup of coffee. The player character rig is contained in 'coffeearm' scene and my test map is 'amaptofinding' scene. If you want to change the reflection position then change the transform of the CoffeeView node3D. Basic up, down, left, right movement controls, and only a single idling animation.